+++ airbyte-cdk/bulk/toolkits/source-tests/src/test/kotlin/io/airbyte/cdk/test/fixtures/cleanup/TestAssetResourceNamerTest.kt
@@ -2,7 +2,7 @@
  * Copyright (c) 2025 Airbyte, Inc., all rights reserved.
  */
 
-package io.airbyte.integrations.sourceTesting
+package io.airbyte.cdk.test.fixtures.cleanup
 
 import io.mockk.every
 import io.mockk.mockk
