+++ airbyte-cdk/bulk/toolkits/source-tests/src/testFixtures/kotlin/io/airbyte/cdk/test/fixtures/cleanup/AssetCleanerShutdownRunner.kt
