+++ airbyte-cdk/bulk/toolkits/source-tests/src/testFixtures/kotlin/io/airbyte/cdk/test/fixtures/connector/JdbcTestDbExecutor.kt
@@ -1,5 +1,5 @@
 /* Copyright (c) 2025 Airbyte, Inc., all rights reserved. */
-package io.airbyte.integrations.sourceTesting
+package io.airbyte.cdk.test.fixtures.connector
 
 import io.airbyte.cdk.command.JdbcSourceConfiguration
 import io.airbyte.cdk.jdbc.JdbcConnectionFactory
