+++ airbyte-cdk/bulk/toolkits/source-tests/src/testFixtures/kotlin/io/airbyte/cdk/test/fixtures/connector/TestDbExecutor.kt
@@ -2,7 +2,7 @@
  * Copyright (c) 2025 Airbyte, Inc., all rights reserved.
  */
 
-package io.airbyte.integrations.sourceTesting
+package io.airbyte.cdk.test.fixtures.connector
 
 interface TestDbExecutor : AutoCloseable {
