+++ airbyte-cdk/bulk/toolkits/source-tests/src/testFixtures/kotlin/io/airbyte/cdk/test/fixtures/tests/FieldTypeMapperTest.kt
@@ -1,5 +1,5 @@
 /* Copyright (c) 2025 Airbyte, Inc., all rights reserved. */
-package io.airbyte.integrations.sourceTesting.tests
+package io.airbyte.cdk.test.fixtures.tests
 
 import com.fasterxml.jackson.databind.JsonNode
 import com.fasterxml.jackson.databind.node.ObjectNode
@@ -9,9 +9,9 @@ import io.airbyte.cdk.data.AirbyteSchemaType
 import io.airbyte.cdk.data.LeafAirbyteSchemaType
 import io.airbyte.cdk.discover.MetaField
 import io.airbyte.cdk.output.BufferingOutputConsumer
+import io.airbyte.cdk.test.fixtures.connector.IntegrationTestOperations
+import io.airbyte.cdk.test.fixtures.connector.TestDbExecutor
 import io.airbyte.cdk.util.Jsons
-import io.airbyte.integrations.sourceTesting.IntegrationTestOperations
-import io.airbyte.integrations.sourceTesting.TestDbExecutor
 import io.airbyte.protocol.models.v0.AirbyteLogMessage
 import io.airbyte.protocol.models.v0.AirbyteMessage
 import io.airbyte.protocol.models.v0.AirbyteRecordMessage
