+++ airbyte-cdk/bulk/toolkits/source-tests/src/testFixtures/kotlin/io/airbyte/cdk/test/fixtures/tests/FullRefreshTest.kt
@@ -1,9 +1,9 @@
 /* Copyright (c) 2025 Airbyte, Inc., all rights reserved. */
 
-package io.airbyte.integrations.sourceTesting.tests
+package io.airbyte.cdk.test.fixtures.tests
 
-import io.airbyte.integrations.sourceTesting.TestAssetResourceNamer
-import io.airbyte.integrations.sourceTesting.TestDbExecutor
+import io.airbyte.cdk.test.fixtures.cleanup.TestAssetResourceNamer
+import io.airbyte.cdk.test.fixtures.connector.TestDbExecutor
 import io.airbyte.protocol.models.v0.SyncMode
 import kotlin.test.assertEquals
 import org.junit.jupiter.api.Assertions.assertTrue
